[KA][COM-89] Convert UI into library that launches with CORE
Made the UI project a library that core can call on so the Editor window launches with Core. Also changed some file structure of IMGUI in the folder and solution files to be more consistent with the rest of the external libraries in the solution. Changed the additional includes and linker directories of UI and Core to work with the new UI library.
