+++ Core/imgui.ini
@@ -1,0 +1,51 @@
+[Window][Viewport]
+Pos=326,27
+Size=708,512
+Collapsed=0
+DockId=0x00000007,0
+
+[Window][Hierarchy]
+Pos=8,27
+Size=316,685
+Collapsed=0
+DockId=0x00000001,0
+
+[Window][Inspector]
+Pos=1036,27
+Size=236,685
+Collapsed=0
+DockId=0x00000004,0
+
+[Window][Browser]
+Pos=326,541
+Size=708,171
+Collapsed=0
+DockId=0x00000006,0
+
+[Window][Console]
+Pos=326,541
+Size=708,171
+Collapsed=0
+DockId=0x00000006,1
+
+[Window][DockSpace Demo]
+Pos=0,19
+Size=1280,701
+Collapsed=0
+
+[Window][Debug##Default]
+Pos=60,60
+Size=400,400
+Collapsed=0
+
+[Docking][Data]
+DockSpace         ID=0xFA06BC56 Window=0x4647B76E Pos=8,27 Size=1264,685 Split=X
+  DockNode        ID=0x00000001 Parent=0xFA06BC56 SizeRef=16,0 Selected=0x29EABFBD
+  DockNode        ID=0x00000002 Parent=0xFA06BC56 SizeRef=48,0 Split=X
+    DockNode      ID=0x00000003 Parent=0x00000002 SizeRef=48,0 Split=Y
+      DockNode    ID=0x00000005 Parent=0x00000003 SizeRef=0,48 Split=Y
+        DockNode  ID=0x00000007 Parent=0x00000005 SizeRef=0,16 Selected=0x13926F0B
+        DockNode  ID=0x00000008 Parent=0x00000005 SizeRef=0,48
+      DockNode    ID=0x00000006 Parent=0x00000003 SizeRef=0,16 Selected=0x49278EEE
+    DockNode      ID=0x00000004 Parent=0x00000002 SizeRef=16,0 Selected=0xE7039252
+
