+++ Core/include/Utils/IDUtils.h
@@ -1,6 +1,12 @@
 #pragma once
+#ifndef IDUTILS_H
+#define IDUTILS_H
+
+
 
 #include "../../External/uuid/uuid_v4.h"
 #include "../../External/uuid/uuid_v4.h"
 
 UUIDv4::UUIDGenerator<std::mt19937_64> uuidGen;
+
+#endif
