+++ UI/UI.vcxproj
@@ -40,13 +40,13 @@
     <CharacterSet>Unicode</CharacterSet>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'" Label="Configuration">
-    <ConfigurationType>Application</ConfigurationType>
+    <ConfigurationType>StaticLibrary</ConfigurationType>
     <UseDebugLibraries>true</UseDebugLibraries>
     <PlatformToolset>v143</PlatformToolset>
     <CharacterSet>Unicode</CharacterSet>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'" Label="Configuration">
-    <ConfigurationType>Application</ConfigurationType>
+    <ConfigurationType>StaticLibrary</ConfigurationType>
     <UseDebugLibraries>false</UseDebugLibraries>
     <PlatformToolset>v143</PlatformToolset>
     <WholeProgramOptimization>true</WholeProgramOptimization>
@@ -109,12 +109,12 @@
       <SDLCheck>true</SDLCheck>
       <PreprocessorDefinitions>_DEBUG;_CONSOLE;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <ConformanceMode>true</ConformanceMode>
-      <AdditionalIncludeDirectories>$(SolutionDir)Core\include\;$(SolutionDir)External\imgui;$(SolutionDir)External\GLFW\include;$(SolutionDir)External\glad\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(SolutionDir)Core\include\;$(SolutionDir)External\imgui\include;$(SolutionDir)External\GLFW\include;$(SolutionDir)External\glad\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
     </ClCompile>
     <Link>
       <SubSystem>Console</SubSystem>
       <GenerateDebugInformation>true</GenerateDebugInformation>
-      <AdditionalLibraryDirectories>..\External\glad\lib;..\External\GLFW\lib;..\External\imgui;%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
+      <AdditionalLibraryDirectories>$(SolutionDir)External\glad\lib;$(SolutionDir)External\GLFW\lib;$(SolutionDir)External\imgui\include</AdditionalLibraryDirectories>
       <AdditionalDependencies>glad.lib;glfw3.lib;opengl32.lib;%(AdditionalDependencies)</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
@@ -126,45 +126,46 @@
       <SDLCheck>true</SDLCheck>
       <PreprocessorDefinitions>NDEBUG;_CONSOLE;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <ConformanceMode>true</ConformanceMode>
-      <AdditionalIncludeDirectories>$(SolutionDir)External\imgui;$(SolutionDir)External\GLFW\include;$(SolutionDir)External\glad\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(SolutionDir)Core\include\;$(SolutionDir)External\imgui\include;$(SolutionDir)External\GLFW\include;$(SolutionDir)External\glad\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
     </ClCompile>
     <Link>
       <SubSystem>Console</SubSystem>
       <EnableCOMDATFolding>true</EnableCOMDATFolding>
       <OptimizeReferences>true</OptimizeReferences>
       <GenerateDebugInformation>true</GenerateDebugInformation>
-      <AdditionalLibraryDirectories>..\External\glad\lib;..\External\GLFW\lib;..\External\imgui;%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
+      <AdditionalLibraryDirectories>$(SolutionDir)External\glad\lib;$(SolutionDir)External\GLFW\lib;$(SolutionDir)External\imgui\include</AdditionalLibraryDirectories>
       <AdditionalDependencies>glad.lib;glfw3.lib;opengl32.lib;%(AdditionalDependencies)</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="..\External\imgui\imgui.cpp" />
-    <ClCompile Include="..\External\imgui\imgui_demo.cpp" />
-    <ClCompile Include="..\External\imgui\imgui_draw.cpp" />
-    <ClCompile Include="..\External\imgui\imgui_impl_glfw.cpp" />
-    <ClCompile Include="..\External\imgui\imgui_impl_opengl3.cpp" />
-    <ClCompile Include="..\External\imgui\imgui_tables.cpp" />
-    <ClCompile Include="..\External\imgui\imgui_widgets.cpp" />
+    <ClCompile Include="..\External\imgui\src\imgui.cpp" />
+    <ClCompile Include="..\External\imgui\src\imgui_demo.cpp" />
+    <ClCompile Include="..\External\imgui\src\imgui_draw.cpp" />
+    <ClCompile Include="..\External\imgui\src\imgui_impl_glfw.cpp" />
+    <ClCompile Include="..\External\imgui\src\imgui_impl_opengl3.cpp" />
+    <ClCompile Include="..\External\imgui\src\imgui_tables.cpp" />
+    <ClCompile Include="..\External\imgui\src\imgui_widgets.cpp" />
     <ClCompile Include="GameComponent.cpp" />
     <ClCompile Include="GameObject.cpp" />
-    <ClCompile Include="main.cpp" />
     <ClCompile Include="src\Browser.cpp" />
     <ClCompile Include="src\Console.cpp" />
     <ClCompile Include="src\Hierarchy.cpp" />
     <ClCompile Include="src\Inspector.cpp" />
+    <ClCompile Include="src\MenuBar.cpp" />
+    <ClCompile Include="src\UI.cpp" />
     <ClCompile Include="src\Viewport.cpp" />
     <ClCompile Include="src\Windows.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\External\imgui\imconfig.h" />
-    <ClInclude Include="..\External\imgui\imgui.h" />
-    <ClInclude Include="..\External\imgui\imgui_impl_glfw.h" />
-    <ClInclude Include="..\External\imgui\imgui_impl_opengl3.h" />
-    <ClInclude Include="..\External\imgui\imgui_impl_opengl3_loader.h" />
-    <ClInclude Include="..\External\imgui\imgui_internal.h" />
-    <ClInclude Include="..\External\imgui\imstb_rectpack.h" />
-    <ClInclude Include="..\External\imgui\imstb_textedit.h" />
-    <ClInclude Include="..\External\imgui\imstb_truetype.h" />
+    <ClInclude Include="..\External\imgui\include\imconfig.h" />
+    <ClInclude Include="..\External\imgui\include\imgui.h" />
+    <ClInclude Include="..\External\imgui\include\imgui_impl_glfw.h" />
+    <ClInclude Include="..\External\imgui\include\imgui_impl_opengl3.h" />
+    <ClInclude Include="..\External\imgui\include\imgui_impl_opengl3_loader.h" />
+    <ClInclude Include="..\External\imgui\include\imgui_internal.h" />
+    <ClInclude Include="..\External\imgui\include\imstb_rectpack.h" />
+    <ClInclude Include="..\External\imgui\include\imstb_textedit.h" />
+    <ClInclude Include="..\External\imgui\include\imstb_truetype.h" />
     <ClInclude Include="GameComponent.h" />
     <ClInclude Include="GameObject.h" />
     <ClInclude Include="include\Browser.h" />
